Skattekort oppdatering og feilretting (#4123)

+++ apps/dolly-backend/src/main/java/no/nav/dolly/bestilling/skattekort/domain/Trekktabell.java
@@ -14,6 +14,6 @@ public class Trekktabell {
     private Trekkode trekkode;
 
     private String tabellnummer;
-    private Integer prosentsats;
-    private Integer antallMaanederForTrekk;
+    private Float prosentsats;
+    private Float antallMaanederForTrekk;
 }
